it's work!!!!!!!!!!!!!!!!!!
finallly the the paint_area function work, hallelujah!

it's check one side, if less than 20000 pixels it paint it
else it check the other side, if him also more than 20000 pixels, it's choose the first one and paint him.

need to change the function that it will paint the smallest one.